api/requirements.txt
mysqlclient==2.0.2
python-dotenv==0.15.0
Django==3.1.4
djangorestframework==3.12.2
django-rest-auth==0.9.5